ttps://www.codewars.com/kata/58ff1c8b13b001a5a50005b4
Write a method that accepts a list of objects of type Animal, and
returns a new list. The new list should be a copy of the original list,
sorted first by the animal's number of legs, and then by its name.

If null is passed, the method should return null. If an empty list is
passed, it should return an empty list back.